Improve WFIRST WFI test suite
Removed get_zernike_coefficient on WavelengthInterpolator as it was mainly for early development and not worth adding test coverage
Altered criteria for autopupil again and wrote tests ensuring it's always selected appropriately for the defined filters
Added checking to the limits of the interpolation ranges in X, Y, and wavelength to raise a RuntimeError if a calculation is started with out-of-bounds values. This will limit WFIRST calculations to 0.76 to 2.0 um.